requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
attrs==21.4.0
    # via pytest
certifi==2022.6.15
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   cryptography
charset-normalizer==2.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via code-annotations
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==6.4.2
    # via pytest-cov
cryptography==37.0.4
    # via
    #   -r requirements/base.txt
    #   social-auth-core
ddt==1.5.0
    # via -r requirements/test.in
defusedxml==0.7.1
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-cas-ng
django-cas-ng==4.3.0
    # via -r requirements/base.txt
idna==3.3
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==1.1.1
    # via pytest
jinja2==3.1.2
    # via code-annotations
lxml==4.9.1
    # via
    #   -r requirements/base.txt
    #   python-cas
markupsafe==2.1.1
    # via jinja2
oauthlib==3.2.0
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via pytest
pbr==5.9.0
    # via stevedore
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pyjwt==2.4.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-cas==1.6.0
    # via
    #   -r requirements/base.txt
    #   django-cas-ng
python-slugify==6.1.2
    # via code-annotations
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2022.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==6.0
    # via code-annotations
requests==2.28.1
    # via
    #   -r requirements/base.txt
    #   python-cas
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements/base.txt
    #   social-auth-core
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   python-cas
social-auth-core==4.1.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
stevedore==4.0.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
urllib3==1.26.10
    # via
    #   -r requirements/base.txt
    #   requests
